requirements.txt
streamlit
streamlit-webrtc
mediapipe
opencv-python-headless
tensorflow-cpu
numpy
aiortc